+++ axelor-account/src/main/java/com/axelor/apps/account/service/move/MoveService.java
@@ -45,7 +45,6 @@ import com.axelor.auth.AuthUtils;
 import com.axelor.auth.db.User;
 import com.axelor.exception.AxelorException;
 import com.google.common.base.Strings;
-import com.google.common.collect.Lists;
 import com.google.inject.Inject;
 import com.google.inject.persist.Transactional;
 
@@ -192,11 +191,7 @@ public class MoveService {
 	public Move createMoveUseInvoiceDue(Invoice invoice) throws AxelorException{
 
 		Company company = invoice.getCompany();
-		Account account = invoice.getPartnerAccount();
-		Partner partner = invoice.getPartner();
-
 		Move move = null;
-
 
 		AccountConfig accountConfig = accountConfigService.getAccountConfig(company);
 
@@ -207,7 +202,7 @@ public class MoveService {
 			MoveLine invoiceCustomerMoveLine = moveToolService.getCustomerMoveLineByLoop(invoice);
 			
 			// Si c'est le même compte sur les trop-perçus et sur la facture, alors on lettre directement
-			if(moveToolService.isSameAccount(debitMoveLines, invoice.getPartnerAccount()))  {
+			if(moveToolService.isSameAccount(debitMoveLines, invoiceCustomerMoveLine.getAccount()))  {
 				List<MoveLine> creditMoveLineList = new ArrayList<MoveLine>();
 				creditMoveLineList.add(invoiceCustomerMoveLine);
 				paymentService.useExcessPaymentOnMoveLines(debitMoveLines, creditMoveLineList);
@@ -219,12 +214,6 @@ public class MoveService {
 
 			// Gestion du passage en 580
 			reconcileService.balanceCredit(invoiceCustomerMoveLine);
-
-			BigDecimal remainingPaidAmount = invoiceCustomerMoveLine.getAmountRemaining();
-			// Si il y a un restant à payer, alors on crée un trop-perçu.
-			if(remainingPaidAmount.compareTo(BigDecimal.ZERO) > 0 )  {
-				this.createExcessMove(invoice, company, partner, account, remainingPaidAmount, invoiceCustomerMoveLine);
-			}
 
 			invoice.setCompanyInTaxTotalRemaining(moveToolService.getInTaxTotalRemaining(invoice));
 		}
@@ -326,55 +315,6 @@ public class MoveService {
 			reconcileService.confirmReconcile(reconcile, true);
 		}
 		return oDmove;
-	}
-
-
-	/**
-	 * Procédure permettant de créer une écriture de trop-perçu
-	 * @param company
-	 * 			Une société
-	 * @param partner
-	 * 			Un tiers payeur
-	 * @param account
-	 * 			Le compte client (411 toujours)
-	 * @param amount
-	 * 			Le montant du trop-perçu
-	 * @param invoiceCustomerMoveLine
-	 * 			La ligne d'écriture client de la facture
-	 * @throws AxelorException
-	 */
-	public void createExcessMove(Invoice refund, Company company, Partner partner, Account account, BigDecimal amount, MoveLine invoiceCustomerMoveLine) throws AxelorException  {
-
-		Journal journal = accountConfigService.getAutoMiscOpeJournal(accountConfigService.getAccountConfig(company));
-
-		Move excessMove = moveCreateService.createMove(journal, company, refund.getCurrency(), partner, null, MoveRepository.TECHNICAL_ORIGIN_AUTOMATIC);
-		excessMove.setInvoice(refund);
-		
-		MoveLine debitMoveLine = moveLineService.createMoveLine(excessMove,
-				partner,
-				account,
-				amount,
-				true,
-				this.today,
-				1,
-				null);
-		excessMove.getMoveLineList().add(debitMoveLine);
-
-		MoveLine creditMoveLine = moveLineService.createMoveLine(excessMove,
-				partner,
-				account,
-				amount,
-				false,
-				this.today,
-				2,
-				null);
-		excessMove.getMoveLineList().add(creditMoveLine);
-
-		moveValidateService.validateMove(excessMove);
-
-		//Création de la réconciliation
-		Reconcile reconcile = reconcileService.createReconcile(debitMoveLine, invoiceCustomerMoveLine, amount, false);
-		reconcileService.confirmReconcile(reconcile, true);
 	}