+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Find kilometric allowance rate per company
 - Remove filter on company for historic timesheet
 - More fixes in expense form
+- Remove the wrong process to create an useless move for excess payment on refund invoice ventilation
+
 
 ## [4.1.2] - 2018-02-05
 ### Improvements